chore: rename file to match class inside
