+++ packages/sshnoports/test/sshnpd_test.dart
@@ -1,4 +1,4 @@
-import 'package:sshnoports/sshnpd/sshnpd_cli_params.dart';
+import 'package:sshnoports/sshnpd/sshnpd_params.dart';
 import 'package:test/test.dart';
 import 'package:args/args.dart';
 import 'package:sshnoports/common/utils.dart';
